Moved validators to a separate file. Moved helpers and validators to a folder 'util'.
