+++ public/_translations/en.json
@@ -1086,6 +1086,7 @@
   "invalid.url": "URL is invalid",
   "invalid.credentials": "Invalid username or password",
   "invalid.doi": "Digital Object Identifier is invalid",
+  "invalid.postalCode": "Postal code is invalid",
   "error.title": "An error occurred",
   "organizations": "Organizations",
   "datasets": "Datasets",
