+++ src/api/dataset.js
@@ -4,7 +4,7 @@ import axiosInstance from './util/axiosInstance';
 import axios_cancelable from './util/axiosCancel';
 import { getOrganization } from './organization';
 import { getInstallation } from './installation';
-import { isUUID } from '../components/helpers';
+import { isUUID } from '../components/util/helpers';
 
 export const searchDatasets = query => {
   return axios_cancelable.get(`/dataset?${qs.stringify(query)}`);
