+++ src/api/node.js
@@ -2,7 +2,7 @@ import qs from 'qs';
 
 import axiosInstance from './util/axiosInstance';
 import axios_cancelable from './util/axiosCancel';
-import { isUUID } from '../components/helpers';
+import { isUUID } from '../components/util/helpers';
 
 export const search = query => {
   return axios_cancelable.get(`/node?${qs.stringify(query)}`);
