+++ src/components/Institution/index.js
@@ -25,7 +25,7 @@ import Exception404 from '../exception/404';
 import { Collections } from './institutionSubtypes';
 import Actions from './institution.actions';
 // Helpers
-import { getSubMenu } from '../helpers';
+import { getSubMenu } from '../util/helpers';
 
 class Institution extends Component {
   constructor(props) {
